+++ packages/core/global.d.ts
@@ -1,0 +1,5 @@
+declare module 'whyframe:app' {
+  export const createApp: (el: HTMLElement) => {
+    destroy: () => void
+  }
+}