+++ packages/core/index.d.ts
@@ -1,11 +1,5 @@
 import type { Plugin } from 'vite'
 
-declare module 'whyframe:app' {
-  export const createApp: (el: HTMLElement) => {
-    destroy: () => void
-  }
-}
-
 type LoadResult = Awaited<ReturnType<NonNullable<Plugin['load']>>>
 
 export interface Attr {