+++ playground/nuxt/global.d.ts
@@ -1,2 +1,1 @@
 /// <reference types="@whyframe/core/global" />
-/// <reference types="astro/client" />