+++ playground/vite-solid/src/vite-env.d.ts
@@ -1,2 +1,2 @@
-/// <reference types="@whyframe/core" />
+/// <reference types="@whyframe/core/global" />
 /// <reference types="vite/client" />