+++ playground/vite-svelte/src/vite-env.d.ts
@@ -1,3 +1,3 @@
-/// <reference types="@whyframe/core" />
+/// <reference types="@whyframe/core/global" />
 /// <reference types="svelte" />
 /// <reference types="vite/client" />