+++ playground/vite-vue/src/vite-env.d.ts
@@ -1,4 +1,4 @@
-/// <reference types="@whyframe/core" />
+/// <reference types="@whyframe/core/global" />
 /// <reference types="vite/client" />
 
 declare module '*.vue' {